requirements.txt
django>=4.2.0
pillow
djangorestframework
djangorestframework-simplejwt 
django-cors-headers
pyjwt
django-jazzmin
django-environ==0.9.0
psycopg2